requirements.txt
arxiv==3.0.0
faiss-cpu==1.13.2
gradio==6.12.0
joblib==1.5.3
matplotlib==3.10.8
numpy==2.4.4
pandas==3.0.2
PyMySQL==1.1.2
python-dotenv==1.2.2
scikit-learn==1.8.0
scipy==1.17.1
SQLAlchemy==2.0.49
torch==2.11.0
tqdm==4.67.3
transformers==5.5.3
